tracing: refactor prometheus exposition, and then some
