+++ cardano-tracer/src/Cardano/Tracer/Utils.hs
@@ -22,8 +22,6 @@ module Cardano.Tracer.Utils
   , initDataPointRequestors
   , initProtocolsBrake
   , logTrace
-  , lift2M
-  , lift3M
   , forMM
   , forMM_
   , nl
@@ -50,13 +48,6 @@ import           Cardano.Tracer.MetaTrace hiding (traceWith)
 import           Cardano.Tracer.Types
 import           Ouroboros.Network.Socket (ConnectionId (..))
 
-
-#if MIN_VERSION_base(4,18,0)
--- Do not know why.
-import           Control.Applicative (liftA3)
-#else
-import           Control.Applicative (liftA2, liftA3)
-#endif
 import           Control.Concurrent (killThread, mkWeakThreadId, myThreadId)
 import           Control.Concurrent.Async (Concurrently(..))
 import           Control.Concurrent.Extra (Lock)
@@ -77,7 +68,6 @@ import           Data.List.Extra (dropPrefix, dropSuffix, replace)
 import qualified Data.Map.Strict as Map
 import qualified Data.Set as S
 import qualified Data.Text as T
-import           Data.Tuple.Extra (uncurry3)
 import           System.IO (hClose, hFlush, stdout)
 import           System.Mem.Weak (deRefWeak)
 import qualified System.Signal as S
@@ -233,14 +223,6 @@ askNodeId TracerEnv{teConnectedNodesNames} nodeName = do
 --   between acceptor's part and forwarder's part will be finished.
 applyBrake :: ProtocolsBrake -> IO ()
 applyBrake stopProtocols = atomically $ modifyTVar' stopProtocols . const $ True
-
--- | Like 'liftM2', but for monadic function.
-lift2M :: Monad m => (a -> b -> m c) -> m a -> m b -> m c
-lift2M f x y = liftA2 (,) x y >>= uncurry f
-
--- | Like 'liftM3', but for monadic function.
-lift3M :: Monad m => (a -> b -> c -> m d) -> m a -> m b -> m c -> m d
-lift3M f x y z = liftA3 (,,) x y z >>= uncurry3 f
 
 forMM :: (Traversable t, Monad m) => m (t a) -> (a -> m b) -> m (t b)
 forMM mVals f = mVals >>= mapM f
